Macspark/marketing-automation/analytics/scripts/requirements.txt
# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Supabase Client
supabase==2.7.4

# Google APIs
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.144.0

# OpenAI
openai==1.51.0

# Data Processing
pandas==2.2.2

# Date/Time
python-dateutil==2.9.0

